Method to load medication from file
Adds in a method (load_medications_and_timers_from_file) to load medications from the listed file, file should store each medication on its own line, with the format name#,#nextDue#,#interval#,#intervalDisplayType where name is a string, nextDue is an integer representing the next time the medication will be due in seconds from the epoch, interval is an integer representation of the amount of seconds between each dose of the medication, and intervalDisplayType is an integer for what type of display the medication should list. the method will return a Dictionary object containing arrays representing each medication.